[Bronze II] Title: 벌집, Time: 0 ms, Memory: 1112 KB -BaekjoonHub

+++ 백준/Bronze/2292. 벌집/벌집.c
@@ -1,14 +1,14 @@
 #include<stdio.h>
 int main()
 {
-	long long int N;
+	int N;
 	int i, count=1;
-	scanf("%lld", &N);
+	scanf("%d", &N);
 	for (i = 0 ;; i++)
 	{
 		if (N <= count + 6*i)
 			break;
 		count+=6*i;
 	}
-	printf("%lld", i+1);
+	printf("%d", i+1);
 }
